* cpplib.c (_cpp_init_stacks): Cast enum for comparison. 	* cppexp.c (lex): Cast enums for comparison. 	* cppinit.c (parse_option): Cast enum for comparison. 	* cpplex.c (cpp_spell_token): Cast enums to int for minus. 	(cpp_output_token): Likewise. 	(cpp_can_paste): Cast enums for comparsion and plus/minus. 	(cpp_avoid_paste): Cast enums for minus and comparison.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@40174 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cppexp.c
@@ -477,7 +477,8 @@ lex (pfile, skip_evaluation, token)
       /* Fall through.  */
 
     default:
-      if ((token->type > CPP_EQ && token->type < CPP_PLUS_EQ)
+      if (((int) token->type > (int) CPP_EQ
+	   && (int) token->type < (int) CPP_PLUS_EQ)
 	  || token->type == CPP_EOF)
 	{
 	  op.op = token->type;
